requirements.txt
amqp==5.1.1
astroid==2.15.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
dill==0.3.6
flower==1.2.0
humanize==4.6.0
isort==5.12.0
kombu==5.2.4
lazy-object-proxy==1.9.0
mccabe==0.7.0
platformdirs==3.2.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pydantic==1.10.7
pylint==2.17.2
pytz==2023.3
redis==4.5.4
six==1.16.0
tomlkit==0.11.7
tornado==6.2
typing_extensions==4.5.0
vine==5.0.0
wcwidth==0.2.6
wrapt==1.15.0
